requirements.txt

beautifulsoup4==4.12.3
discord==2.3.2
discord.py==2.3.2
python-dotenv==1.0.1
requests==2.32.3
